requirements.txt
argparse==1.2.1
awesome-slugify==1.6.5
Django==1.10
django-braces==1.9.0
django-groups-manager==0.4.1
django-guardian==1.4.5
django-mptt==0.8.5
django-oauth-toolkit==0.10.0
djangorestframework==3.4.5
httpie==0.9.4
jsonfield==1.0.3
mysqlclient==1.3.7
oauthlib==1.0.3
pip==1.5.4
Pygments==2.1.3
python-memcached==1.58
regex==2016.07.21
requests==2.10.0
setuptools==2.2
six==1.10.0
Unidecode==0.04.19
wsgiref==0.1.2
